LayoutTests/fast/page-color-sampling/page-color-sampling-skips-small-elements.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=100 obscuredInset.left=100 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body, html {
            width: 100%;
            margin: 0;
            font-family: system-ui;
        }

        .left {
            position: fixed;
            top: calc(50% - 10px);
            height: 100px;
            width: 20px;
            background: red;
            border-top-right-radius: 6px;
            border-bottom-right-radius: 6px;
        }

        .top {
            position: fixed;
            top: 0;
            left: calc(50% - 50px);
            width: 100px;
            height: 20px;
            background: red;
            border-bottom-left-radius: 6px;
            border-bottom-right-radius: 6px;
        }

        .tall {
            width: 10px;
            height: 5000px;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.ensurePresentationUpdate();
        edgeColors = await UIHelper.fixedContainerEdgeColors();
        shouldBeNull("edgeColors.top");
        shouldBeNull("edgeColors.left");
        finishJSTest();
    });
    </script>
</head>
<body>
<div class="top"></div>
<div class="left"></div>
<div class="tall"></div>
</body>
</html>
