LayoutTests/fast/page-color-sampling/page-color-sampling-skips-small-elements-expected.txt

PASS edgeColors.top is null
PASS edgeColors.left is null
PASS successfullyParsed is true

TEST COMPLETE
